sortArrowIcon and sortAnimationDuration tests pass
